civicxchain-platform/civicXchain-Database/satellite_apis.py
import requests
from datetime import datetime
import os
from database_manager import DatabaseManager

def connect_to_nasa_api(lat, lon, date_range):
    """Real NASA MODIS API integration"""
    
    # You'll need to register at https://earthdata.nasa.gov/
    nasa_api_key = os.getenv("NASA_API_KEY")  # Add this to your environment
    
    # Real NASA MODIS endpoint
    api_url = "https://modis.ornl.gov/rst/api/v1/MOD13Q1/subset"
    
    params = {
        "latitude": lat,
        "longitude": lon, 
        "startDate": date_range[0],
        "endDate": date_range[1],
        "kmAboveBelow": 0,
        "kmLeftRight": 0
    }
    
    headers = {
        "Authorization": f"Bearer {nasa_api_key}"
    }
    
    try:
        response = requests.get(api_url, params=params, headers=headers)
        response.raise_for_status()
        return response.json()
    except requests.exceptions.RequestException as e:
        print(f"❌ NASA API Error: {e}")
        return None

def get_air_quality_data(lat, lon):
    """Connect to OpenAQ API for air quality"""
    
    openaq_url = "https://api.openaq.org/v2/measurements"
    params = {
        "coordinates": f"{lat},{lon}",
        "radius": 10000,  # 10km radius
        "limit": 100,
        "order_by": "datetime",
        "sort": "desc"
    }
    
    try:
        response = requests.get(openaq_url, params=params)
        response.raise_for_status()
        return response.json()
    except Exception as e:
        print(f"❌ OpenAQ API Error: {e}")
        return None

def get_weather_data(lat, lon):
    """Connect to OpenWeather API"""
    
    weather_api_key = os.getenv("OPENWEATHER_API_KEY")
    weather_url = f"https://api.openweathermap.org/data/2.5/weather"
    
    params = {
        "lat": lat,
        "lon": lon,
        "appid": weather_api_key,
        "units": "metric"
    }
    
    try:
        response = requests.get(weather_url, params=params)
        response.raise_for_status()
        return response.json()
    except Exception as e:
        print(f"❌ Weather API Error: {e}")
        return None

def analyze_forest_cover_change(satellite_data):
    """Process real satellite imagery for forest cover analysis"""
    
    # This would typically use libraries like:
    # - rasterio for satellite image processing
    # - numpy for numerical analysis
    # - sklearn for machine learning classification
    
    try:
        # Example processing (you'd need actual image processing here)
        ndvi_values = satellite_data.get('NDVI', [])
        
        # NDVI > 0.3 typically indicates vegetation
        forest_pixels = [v for v in ndvi_values if v > 0.3]
        total_pixels = len(ndvi_values)
        
        if total_pixels > 0:
            forest_percentage = (len(forest_pixels) / total_pixels) * 100
            return {
                "forest_cover_percentage": round(forest_percentage, 2),
                "confidence": 0.85,  # Based on NDVI analysis quality
                "analysis_method": "NDVI_threshold",
                "total_pixels_analyzed": total_pixels
            }
        else:
            return None
            
    except Exception as e:
        print(f"❌ Forest analysis error: {e}")
        return None