civicxchain-platform/civicXchain-Database/requirements.txt
# PostgreSQL Database Requirements
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0

# Optional: For async PostgreSQL operations
asyncpg>=0.25.0
